requirements.txt
firebase-admin
streamlit
streamlit-authenticator
PyYAML
python-dotenv
faiss-cpu
numpy
scikit-learn